refactor: コマンドを自己完結モジュール化し registry で集約
理想ディレクトリ構造への移行 Phase 4。挙動変更はなく、コマンドの持ち方を変える。

これまで新しいコマンドを1つ追加するには3箇所の編集が必要だった:
  1. config/constants/commands.ts に表示名を追加
  2. register.ts (1526行) に SlashCommandBuilder の定義を追加
  3. command_handler.ts の if-else 連鎖 (34分岐) に分岐を追加
どれか1つ忘れても型エラーにはならず、「登録したのにハンドラがない」
「ハンドラはあるのに登録されない」が起こりうる構造だった。

本コミットで以下に変更した:
- registry/types.ts: コマンドは定義と実行関数をセットで持つ、という型を定義。
- features/<機能>/commands/*.ts: 37コマンドを1コマンド=1ファイルに分解。
  定義(register.ts から逐語的に移設)と execute を同じファイルに置く。
  コマンド名は commandNames の参照をやめ、定義に直接インライン化した。
- registry/command_registry.ts: 全コマンドを明示 import で集約。
  この配列の順序がそのまま Discord への登録順になる。
- registry/register.ts: commandDefinitions を registry から生成。
- command_handler.ts / context_handler.ts: if-else 連鎖を廃止し、
  registry から生成した Map による dispatch に置き換え。
  guildOnly の区別・エラー時の挙動・ボイスロックの応答済みガードは維持。

削除: register.ts、config/constants/commands.ts、command_builder.ts × 2。

これによりコマンド追加は「commands/ にファイルを1つ作り、registry の配列に
1行足す」だけになり、登録漏れ・dispatch 漏れが構造的に起きなくなった。

挙動不変の裏付け: commandDefinitions のスナップショットテストが無修正で一致
(37定義の JSON 表現と登録順が完全に同一)。全80テスト green。
