+++ src/app/handlers/context_handler.ts
@@ -1,13 +1,18 @@
 import { CacheType, MessageContextMenuCommandInteraction } from 'discord.js';
 
-import { commandNames } from '@/config/constants/commands';
-import { buttonEnable } from '@/features/admin/button_enabler/enable_button';
-import { createRecruitEditor } from '@/features/recruit/interactions/edit_recruit/recruit_editor';
 import { sendCommandLog } from '@/infra/logging/command_log';
 import { log4js_obj } from '@/infra/logging/log4js';
 import { sendErrorLogs } from '@/infra/logging/send_error_logs';
+import { messageContextMenuCommands } from '@/registry/command_registry';
+import { exists } from '@/shared/assert';
+
+import type { MessageContextMenuCommand } from '@/registry/types';
 
 const logger = log4js_obj.getLogger('interaction');
+
+const contextMenuCommands = new Map<string, MessageContextMenuCommand>(
+    messageContextMenuCommands.map((command) => [command.definition.name, command]),
+);
 
 export async function call(interaction: MessageContextMenuCommandInteraction<CacheType>) {
     try {
@@ -17,10 +22,9 @@ export async function call(interaction: MessageContextMenuCommandInteraction<Cac
     }
 
     if (interaction.inGuild()) {
-        if (interaction.commandName === commandNames.buttonEnabler) {
-            await buttonEnable(interaction);
-        } else if (interaction.commandName === commandNames.recruitEditor) {
-            await createRecruitEditor(interaction);
+        const command = contextMenuCommands.get(interaction.commandName);
+        if (exists(command)) {
+            await command.execute(interaction);
         }
     }
     return;
