--- src/config/constants/commands.ts
@@ -1,39 +1,0 @@
-export const commandNames = {
-    shutdown: 'shutdown',
-    vclock: 'ボイスロック',
-    close: 'close',
-    friend_code: 'friend_code',
-    wiki: 'wiki',
-    kansen: 'kansen',
-    team_divider: 'チーム分け',
-    timer: 'timer',
-    pick: 'pick',
-    voice_pick: 'vpick',
-    buki: 'buki',
-    show: 'show',
-    help: 'help',
-    ban: 'ban',
-    ch_manager: 'ch_management',
-    voice: 'voice',
-    experience: 'イカ部歴',
-    other_game: '別ゲー募集',
-    regular: 'ナワバリ募集',
-    event: 'イベマ募集',
-    anarchy: 'バンカラ募集',
-    salmon: 'サーモンラン募集',
-    raiders: 'レイダース募集',
-    fesA: 'フウカ陣営',
-    fesB: 'マンタロー陣営',
-    fesC: 'ウツホ陣営',
-    private: 'プラベ募集',
-    buttonRecruit: '募集ボタン',
-    buttonEnabler: 'ボタンを有効化する',
-    recruitEditor: '募集を編集',
-    voiceChannelMention: 'ボイスメンション',
-    channelSetting: 'チャンネル設定',
-    uniqueChannelSetting: '固有チャンネル設定',
-    uniqueRoleSetting: '固有ロール設定',
-    variablesSettings: '環境変数設定',
-    joinedDateFixer: '入部日修正',
-    festivalSettings: 'フェスカテゴリ設定',
-};
