+++ src/features/admin/commands/button_enabler.ts
@@ -1,0 +1,16 @@
+import { ApplicationCommandType, ContextMenuCommandBuilder } from 'discord.js';
+
+import { buttonEnable } from '@/features/admin/button_enabler/enable_button';
+
+import type { MessageContextMenuCommand } from '@/registry/types';
+
+const buttonEnabler = new ContextMenuCommandBuilder()
+    .setName('ボタンを有効化する')
+    .setType(ApplicationCommandType.Message)
+    .setDMPermission(false);
+
+export const buttonEnablerCommand: MessageContextMenuCommand = {
+    kind: 'contextMenu',
+    definition: buttonEnabler,
+    execute: buttonEnable,
+};
