+++ src/features/recruit/commands/recruit_editor.ts
@@ -1,0 +1,16 @@
+import { ApplicationCommandType, ContextMenuCommandBuilder } from 'discord.js';
+
+import { createRecruitEditor } from '@/features/recruit/interactions/edit_recruit/recruit_editor';
+
+import type { MessageContextMenuCommand } from '@/registry/types';
+
+const recruitEditor = new ContextMenuCommandBuilder()
+    .setName('募集を編集')
+    .setType(ApplicationCommandType.Message)
+    .setDMPermission(false);
+
+export const recruitEditorCommand: MessageContextMenuCommand = {
+    kind: 'contextMenu',
+    definition: recruitEditor,
+    execute: createRecruitEditor,
+};
