+++ src/features/utils/commands/experience.ts
@@ -1,0 +1,17 @@
+import { SlashCommandBuilder } from 'discord.js';
+
+import { handleIkabuExperience } from '@/features/utils/other/experience';
+
+import type { GuildChatInputCommand } from '@/registry/types';
+
+const experience = new SlashCommandBuilder()
+    .setName('イカ部歴')
+    .setDescription('イカ部歴を表示します。')
+    .setDMPermission(false);
+
+export const experienceCommand: GuildChatInputCommand = {
+    kind: 'chatInput',
+    guildOnly: true,
+    definition: experience,
+    execute: handleIkabuExperience,
+};
